requirements.txt
python-dotenv
google-generativeai
beautifulsoup4
selenium
webdriver-manager